Parciales/parcial2/3-envios-service-python/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
SQLAlchemy[asyncio]
asyncmy         
grpcio
python-dotenv
protobuf